Move cache code to the cache module
Move the cache related code under the thunderstore.cache module
