+++ django/thunderstore/community/models/package_listing.py
@@ -3,7 +3,7 @@ from django.db.models import Q, signals
 from django.urls import reverse
 from django.utils.functional import cached_property
 
-from thunderstore.core.cache import CacheBustCondition, invalidate_cache
+from thunderstore.cache.cache import CacheBustCondition, invalidate_cache
 from thunderstore.core.mixins import TimestampMixin
