+++ django/thunderstore/repository/api/v1/viewsets.py
@@ -8,7 +8,7 @@ from rest_framework.generics import get_object_or_404
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
-from thunderstore.core.cache import BackgroundUpdatedCacheMixin
+from thunderstore.cache.cache import BackgroundUpdatedCacheMixin
 from thunderstore.core.utils import CommunitySiteSerializerContext
 from thunderstore.repository.api.v1.serializers import PackageListingSerializer
 from thunderstore.repository.cache import get_package_listing_queryset
