+++ django/thunderstore/repository/cache.py
@@ -1,5 +1,5 @@
+from thunderstore.cache.cache import CacheBustCondition, cache_function_result
 from thunderstore.community.models import CommunitySite, PackageListing, Q
-from thunderstore.core.cache import CacheBustCondition, cache_function_result
 
 
 @cache_function_result(cache_until=CacheBustCondition.any_package_updated)
